Improve finalizers to receive exceptions.
Now finalizers are represented as `Option[Throwable] => Unit`. If an
exception terminates a running pipe, the finalizers receive the
exception. Note that _all_ of them receive the exception, not just those
from the failed component. Finalizers can use the exception to decide
how severe is the situation (i.e. `Exception` vs `VirtualMachineError`)
and act accordingly.

Finalizers can also be run by users for convenience. In such a case,
they receive `None` instead of an exception.